# apps/api-server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Nexus Platform API
  description: |
    API REST da plataforma Nexus — Gestão em Saúde & Compliance LGPD.

    ## Autenticação
    A API usa **JWT Bearer Token**. Faça login em `POST /auth/login` para obter o token
    e inclua-o no header `Authorization: Bearer <token>` em todas as requisições protegidas.

    ## Multi-Tenant
    Todos os dados são isolados por `company_id`. O token JWT carrega o `company_id` do
    usuário autenticado e o banco filtra automaticamente — nenhum dado de outro tenant
    é acessível.

    ## Módulos
    Alguns endpoints exigem módulos específicos ativos no tenant:
    - `/consents/*` → requer módulo `NEXUSLEGAL`

  version: 1.0.0
  contact:
    name: Nexus Platform
    email: dev@nexus.app
servers:
  - url: https://nexus-platform-x1r5.onrender.com
    description: Produção (Render)
  - url: http://localhost:3001
    description: Desenvolvimento local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtido via `POST /api/v1/auth/login`.
        Expira em 15 minutos — use o refresh token para renovar.
  schemas:
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cnpj:
          type: string
          nullable: true
        nome_fantasia:
          type: string
        tenant_type:
          type: string
          enum: [MED, CLIN, ODONTO, LAB, IMG, ADM]
        status:
          type: string
          enum: [ACTIVE, TRIAL, BLOCKED, CANCELLED]
        active_modules:
          type: array
          items:
            type: string
        user_count:
          type: integer
        trial_expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    CreateTenantRequest:
      type: object
      required: [nome_fantasia, tenant_type]
      properties:
        nome_fantasia:
          type: string
          example: Clínica São Lucas
        cnpj:
          type: string
          example: '00.000.000/0001-00'
        razao_social:
          type: string
        tenant_type:
          type: string
          enum: [MED, CLIN, ODONTO, LAB, IMG, ADM]
        active_modules:
          type: array
          items:
            type: string
          example: [NEXUSMED, NEXUSLEGAL]
        trial_days:
          type: integer
          default: 30
    PlatformStats:
      type: object
      properties:
        total_tenants:
          type: integer
        active_tenants:
          type: integer
        trial_tenants:
          type: integer
        blocked_tenants:
          type: integer
        cancelled_tenants:
          type: integer
        new_last_30d:
          type: integer
        total_users:
          type: integer

# ── Paths ─────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/root/stats:
    get:
      tags: [Root]
      summary: Métricas globais da plataforma
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Métricas
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PlatformStats'
  /api/v1/root/tenants:
    get:
      tags: [Root]
      summary: Listar todos os tenants
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de tenants
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
                  total:
                    type: integer

    post:
      tags: [Root]
      summary: Provisionar novo tenant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant provisionado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tenant'
  /api/v1/root/tenants/{id}/status:
    patch:
      tags: [Root]
      summary: Alterar status do tenant
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [ACTIVE, BLOCKED, TRIAL, CANCELLED]
      responses:
        '200':
          description: Status atualizado
        '404':
          description: Tenant não encontrado
  /api/v1/root/tenants/{id}/modules:
    patch:
      tags: [Root]
      summary: Atualizar módulos ativos do tenant
      description: Controla o licenciamento — quais módulos o tenant tem acesso.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [active_modules]
              properties:
                active_modules:
                  type: array
                  items:
                    type: string
                  example: [NEXUSMED, NEXUSLEGAL, NEXUSADM]
      responses:
        '200':
          description: Módulos atualizados
        '400':
          description: Módulos inválidos
  /api/v1/root/tenants/{id}:
    delete:
      tags: [Root]
      summary: Cancelar tenant (soft delete)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant cancelado
        '404':
          description: Tenant não encontrado